# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple app when learning Go",
        "title": "My Go App API",
        "contact": {
            "name": "Tony Li",
            "url": "https://github.com/tonyltf",
            "email": "tingfung.tony@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
        "/healthcheck": {
            "get": {
                "summary": "Health checking",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/price/{exchange_pair}": {
            "get": {
                "summary": "Get the exchange price",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Currency Pair",
                        "name": "exchange_pair",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "timestamp",
                        "name": "timestamp",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "number"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/price/{exchange_pair}/average": {
            "get": {
                "summary": "Get the average exchange price",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Currency Pair",
                        "name": "exchange_pair",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "From time",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "To time",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "number"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}